requirements.txt
requests>=2.31
pandas>=2.1
sqlalchemy>=2.0
lxml>=4.9
openpyxl>=3.1
xlrd>=2.0
tqdm>=4.66
python-dotenv>=1.0
apscheduler>=3.10
beautifulsoup4>=4.12
html5lib>=1.1
streamlit>=1.36
plotly>=5.20
